requirements.txt
streamlit
openai
python-dotenv
fastapi
uvicorn

